x/computing/keeper/msg_server_devices_test.go
package keeper_test

import (
	"testing"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"

	"exabits/x/computing/types"
)

func TestDevicesMsgServerCreate(t *testing.T) {
	srv, ctx := setupMsgServer(t)
	creator := "A"
	for i := 0; i < 5; i++ {
		resp, err := srv.CreateDevices(ctx, &types.MsgCreateDevices{Creator: creator})
		require.NoError(t, err)
		require.Equal(t, i, int(resp.Id))
	}
}

func TestDevicesMsgServerUpdate(t *testing.T) {
	creator := "A"

	for _, tc := range []struct {
		desc    string
		request *types.MsgUpdateDevices
		err     error
	}{
		{
			desc:    "Completed",
			request: &types.MsgUpdateDevices{Creator: creator},
		},
		{
			desc:    "Unauthorized",
			request: &types.MsgUpdateDevices{Creator: "B"},
			err:     sdkerrors.ErrUnauthorized,
		},
		{
			desc:    "Unauthorized",
			request: &types.MsgUpdateDevices{Creator: creator, Id: 10},
			err:     sdkerrors.ErrKeyNotFound,
		},
	} {
		t.Run(tc.desc, func(t *testing.T) {
			srv, ctx := setupMsgServer(t)
			_, err := srv.CreateDevices(ctx, &types.MsgCreateDevices{Creator: creator})
			require.NoError(t, err)

			_, err = srv.UpdateDevices(ctx, tc.request)
			if tc.err != nil {
				require.ErrorIs(t, err, tc.err)
			} else {
				require.NoError(t, err)
			}
		})
	}
}

func TestDevicesMsgServerDelete(t *testing.T) {
	creator := "A"

	for _, tc := range []struct {
		desc    string
		request *types.MsgDeleteDevices
		err     error
	}{
		{
			desc:    "Completed",
			request: &types.MsgDeleteDevices{Creator: creator},
		},
		{
			desc:    "Unauthorized",
			request: &types.MsgDeleteDevices{Creator: "B"},
			err:     sdkerrors.ErrUnauthorized,
		},
		{
			desc:    "KeyNotFound",
			request: &types.MsgDeleteDevices{Creator: creator, Id: 10},
			err:     sdkerrors.ErrKeyNotFound,
		},
	} {
		t.Run(tc.desc, func(t *testing.T) {
			srv, ctx := setupMsgServer(t)

			_, err := srv.CreateDevices(ctx, &types.MsgCreateDevices{Creator: creator})
			require.NoError(t, err)
			_, err = srv.DeleteDevices(ctx, tc.request)
			if tc.err != nil {
				require.ErrorIs(t, err, tc.err)
			} else {
				require.NoError(t, err)
			}
		})
	}
}
